Adding automatic cleanup of output dir and scoping to the run
The edX pipelines end up using the same directory location when they execute at the same time, leading to resource contention, failed pipelines, and likely inaccurate data. This adds automatic scoping of the output directory to the run id to avoid that contention. It also changes their schedule to be offset from each other and removes the directory on pipeline completion.