refactor,feat: Restructure the repository to support data lakehouse
We are working on building out a data platform with the core building block being a data
lakehouse that is populated and managed with an ELT workflow. To that end this adds
additional directory structures, and renames others, to make it clearer which parts of
the code are for which task.

- Rename ol_data_pipelines -> ol_orchestrate to make it clear that this code is relevant
to data orchestration workflows
- Create an ol_dbt directory for holding dbt model definitions and associated
configurations and workflows
- Add Pants to the repository for handling build and packaging of Dagster orchestration
pipelines to simplify build and deployment workflows