Restructuring to use recommended Dagster project layout and add Pants
In order to reduce the overhead of building and deploying the different pipelines this
updates the structure of the repository and build flow.

The goal is to have BUILD targets defined that specify the Dagster repository to
package, build that into a Python distribution, and install that distribution into a
Docker image that will get published for deployment.

This also moves to a layout with ops/jobs/graphs as the primary concern so that we can
abstract the actual tasks across the different business concerns to improve logic reuse.