feat: Add build logic for Dagster container images
As part of the updated build/deployment we want to have separate images for the Dagit
and dagaster-daemon processes, which are also separate from the user pipeline code so
that they can all be built, deployed, and scaled independently. For the user pipelines
we also want to ensure that the dbt project is available in the runtime
environment. This does the following:
- Copy all files related to the dbt project into user pipeline images by default
- Create a multi-stage build for Dagit/dagster-daemon to avoid duplicate logic
- Moves the Dagster-specific workspace and Dagster yaml files into the `ol_orchestrate` directory
- Moves the dbt project files to the proper directory level in the repo
- Adds the initial work to package up collections of Dagster pipelines based on the 'repository' as the entry-point for the Python distribution