fix,refactor: Condense and simplify Dagster pipeline/container build
The Pants build configuration was starting to become more complicated which led to
conflicts of module ownership that were non-trivial to resolve and didn't provide any
functional benefit. This condenses the builds to a single package and relies on package
imports in the docker-compose to run the proper pipelines at deploy time.