fix: Reduce scope of assets for reconciliation sensor
Trying to execute the sensor against all of the assets rendered in the ELT graph is too
slow. This reduces the scope to only watch for assets in the dbt intermediate and mart
groupings so that it can trigger more consistently. This also updates the
materialization method in dbt to be tables since we don't have the models configured
properly to actually run incrementally.