refactor: Migrate to use Ruff for linting
The wemake-python-styleguide project appears to be inactive and has an unpredictable
future. In its current state it is incompatible with the newest Flake8 version, which
prevents us from upgrading our pre-commit hooks. Ruff is a tool that subsumes a number
of our other pre-commit checks and does it much faster, while providing a substantial
rule set that largely matches our previous rules.