fix: Update dbt configs to work with new targets
The namespaced schemas with the new targets work, but cause errors in the `source.yml`
files because those were relying on `target.name` which resulted in invalid schema
names. This updates those configs to use the `target.database` and a sanitized
`target.schema` to ensure the proper locations are being used.