fix: Resolve startup failure in dbt pipeline
The dbt pipeline was failing to start up due to the addition of the new targets that
required a var for interpolation. This adds a default value for that var. In addition it
switches to using the dbt manifest.json for loading the project structure to improve
startup speeds. That also means generating the manifest file as part of the dbt
container build process.

Lastly, this updates the Python and Pants version specifiers to stay up to date with
recent releases. With the newer versions of Pants we no longer need to be constrained by
the version that Pants needs to be able to bootstrap itself, allowing us to instead
track the Python version that we want to support in our code.