fix: Use built-in user/pass auth for Airbyte
Airbyte recently added basic auth to their default OSS deployment and so Dagster exposed
the user/pass parameters as inputs to the resource object. This removes the
HTTPBasicAuth object passed to requests in favor of the supported parameters. This is
also necessary because the HTTPBasicAuth object was trying to be JSON serialized in the
gRPC response for the running container, preventing it from being loaded by Dagit.