737 Log Normalization Dagster Pipeline (#750)
* Added a new Dagster pipeline to normalize historical tracking log data.
Added new ops to:
-handle getting S3 bucket prefixes and objects.
-load S3 objects directly to a table in a DuckDB instance.
-transform records in DuckDB with SQL updates.
-write normalized files back to S3
Added normalize_tracking_logs job to string together the new ops
Started writing up the repository for the new pipeline.
TODO: Fix eslint errors, properly import/configure a duckDB resource, finalize the config_schema for the new pipeline

* Autoformat updates.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* refactor: String together ops and daily partition

- Add a daily partition definition for being able to launch backfills
- Wire together the ops and pass through the log date
- Add definition for DuckDB resource

* refactor: Remove DuckDB IO Manager in favor of persisted DB

Just using the DuckDB Resource allows for passing the data between ops through the
shared DB file. This refactors the ops for normalizing tracking logs to string together
via out/in `Nothing` types. This also handles passing the IAM credentials for use in DuckDB

* config: Add build definitions for packaging log processing pipeline

The pipeline needs to be packaged in a Docker image. This adds the build metadata to be
able to wire it into a container to be run in QA/Production

* Delete normalize_logs.py

* "Added outfile logic using duckdb sql queries
Refactored transform op to alter and update all columns dynamically"

* "Refactored outfile logic to get code working"

* Create job definitions for normalizing logs in each deployment

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tobias Macey <tmacey@mit.edu>