config: Use production raw data for QA environment
The QA data in the warehouse is not useful for verifying code changes against. This
updates the dbt profile configuration to create a QA namespaced schema for building the
tables in during builds in the QA environment, so that we can use that for properly
validating changes before promoting to production.