fix: Update signature of Pydantic Field objects
Dagster had a concept of a `Field` that had a specific signature for managing the type
information, default values, and whether a field was required. With the new Pydantic
interface that is managed through the attribute type annotations and the Pydantic
`Field` object which has a different signature. This updates the usage of those `Field`
instances to match the signature for Pydantic.