fix,refactor: Migrate GCS sensor to use Pythonic config and fix sensor
The sensor code was written with a hard-coded `run_key` which would always cause it to
skip any runs after the first because that run key had already been triggered. This
migrates the GCP GCS resource to use the Pythonic config so it can be more easily
integrated into the sensor, and then updates the usage of the cursor to do set
arithmetic for any files that have not yet been processed.