fix: Ensure that all edX files get uploaded
Because the logic is iterating over the ResultsDir resource, there is no guarantee that
the files are actually present in the resource as created in that step. This changes to
accept a list of DagsterPath objects that get iterated over and recursively uploaded to
be more explicit about passing of files.