Pipeline for Open edX course structure documents (#881)
* feat: Add operation to fetch course structures from API

This refactors the edX client to be a Pythonic resource class to reduce duplicate code
and simplify usage. It also implements the logic necessary to fetch the course structure
documents via API and write them to disk as newline delimited JSON.

* feat: Add graph definition for course structure uploads

The existing Open edX pipelines are intended for external data consumers. This adds a
new graph definition focused on data extraction that is needed to power the MIT OL data platform.

* config,feat: Add automatic secret retrieval in config function

- Add pydantic-vault dependency
- Add config function for course structure job using pydantic-vault object

* feat: Structure course structure document for use as a table

Add the course ID and retrieval time as attributes in the JSON document so that when it
is ingested into the data platform a table it can be used as a reference column.

* fix,feat: Add automatic update of access token when expired

The JWT access token issued by Open edX has a default expiration of 1 hour. This
automatically re-fetches a token when the one retrieved expires. This will resolve
failures associated with long-running pipelines targeting Open edX

* refactor: Migrate open edX pipelines to use new resource

- Use the new Open edX API resource
- Use pydantic-vault for managing RunConfig instead of relying on yaml files on disk

* feat: Add nightly schedule for open edx extract jobs for OL