refactor: Rename GCS and S3 sensors to signal how they operate
As we start investing in asset-based pipeines we will want to have a few different
flavors of object storage sensors. The current ones trigger a single run for a block of
files. We will want to have an option for triggering one run per file, as well as
generating one or more asset materializations for files in object storage.