feat: Add external asset sensor for edxorg data archives
This adds a sensor definition that populates a dynamic partition definition with the
archive files as they exist in GCS. This will allow us to use the software defined
assets machinery to more seamlessly process the edxorg data as it arrives. It also adds
flexibility for using the various downstream assets produced for a wider variety of
operations, including indexing into the search context for MIT Open.