Edxorg archive processing (#991)
* doc: Add documentation of edx.org data archives processing

Add a module doc for the pipeline definitions file that gives details of the edx.org
data archives format and how we process them.

* refactor: Extract course structure writing to shared method

We will need to be able to write the processed course structure documents from edx.org
data in a format that matches our open edx data.

* refactor: Use Vault for retrieving GCS credentials

* refactor: Rename GCS and S3 sensors to signal how they operate

As we start investing in asset-based pipeines we will want to have a few different
flavors of object storage sensors. The current ones trigger a single run for a block of
files. We will want to have an option for triggering one run per file, as well as
generating one or more asset materializations for files in object storage.

* feat: Add external asset sensor for edxorg data archives

This adds a sensor definition that populates a dynamic partition definition with the
archive files as they exist in GCS. This will allow us to use the software defined
assets machinery to more seamlessly process the edxorg data as it arrives. It also adds
flexibility for using the various downstream assets produced for a wider variety of
operations, including indexing into the search context for MIT Open.

* feat: Add multi-asset definition for edxorg archive contents

There are numerous sub-assets of the edxorg archive that get generated. This models
those assets and categorizes them by parsing the file paths.

* feat: Load edxorg archive from GCS with IOmanager

- Add an IO Manager for loading files from GCS
- Add an IO Manager for handling file up/down loads for GCS and S3
- Process archive files and sub-files

* feat: Materialize sub-assets of edxorg archive file to S3

* refactor: Minor cleanup of extraneous debug logging

* feat: Process edxorg tracking logs as software defined assets

* feat: Add processing of edx.org course structure documents

* config: Switch to use docker run launcher

Configure Dagster to use the Docker run launcher and load the code locations as Python
modules to simplify the operational aspects of adding new pipeines.