fix: Allow for intermittent http status errors to not fail course exports
Any http errors in the bulk course export for legacy edX pipeline causes the whole thing
to fail, even though subsequent tries are likely to return a valid response. This adds a
try/except for that case and allows for continuing to cycle through the active tasks.

This also removes the unused asset materialization events to remove the visual clutter
in the assets page.