Fix issues with Dagster local development (#1479)
* fix: Decode stdout bytes from mongodump as text

* fix: Resolve build failures for local docker-compose

Trying to launch the local docker-compose stack fails due to trying to install the
ol_orchestrate package. This adds a `--no-root` command to Poetry to address that error.

* fix: Resolve GitHub token auth for Vault client

The client object in the resource was constantly being recreated on every access. This
adds a conditional initialization and otherwise returns the initialized object.

* feat: Use Postgres for Dagster storage

* feat: Migrate to UV for faster and cleaner builds

* fix: Resolve errors in mongodump command for legacy edX export

* fix: Set default sensor statuses to stopped

When launching a local Dagster we don't want all of the sensors to automatically launch,
because that will kick off a lot of compute that isn't needed unless that is what is
being tested.