chore: Restructure dagster projects to use dg-style projects under a workspace
The newer versions of Dagster are moving to a more opinionated structure oriented around
the use of the `dg` CLI to make it easier to manage local development, as well as
leaning into their new `Components` abstraction that makes it easier to stamp out new
data assets. This refactors our current project to use this new layout. The majority of
the work was done with GH Copilot CLI using Claude Sonnet 4.5