chore: Restructure dagster projects to use dg-style projects (#1710)
* chore: Restructure dagster projects to use dg-style projects under a workspace

The newer versions of Dagster are moving to a more opinionated structure oriented around
the use of the `dg` CLI to make it easier to manage local development, as well as
leaning into their new `Components` abstraction that makes it easier to stamp out new
data assets. This refactors our current project to use this new layout. The majority of
the work was done with GH Copilot CLI using Claude Sonnet 4.5

* chore: Move shared library to new location and fix code location loading

* chore: Remove un-referenced repository files

* Move code locations to top-level dg_projects

* fix: Properly load multiple openedx deployments as scoped components

* chore: Scope dependencies to specific code locations

* Update dockerfiles to build properly

* feat: Add script for managing dependencies and venvs of dg projects

* fix: Resolve Dockerfile builds for Dagster entrypoint

* fix: Update docker builds to work properly when launched

Updated to not install the ol-orchestrate-lib as editable when building images for deployment.

* config: Let docker compose work for local development

* fix uv lock

* Move edxorg course bundle sensor to proper dg project

* Pull in latest canvas changes

* chore: Prune ol-orchestrate-lib of single-use modules

Move the modules that are only used in a single code location to the appropriate
dg_project and udpate third-party dependencies accordingly.

* Addres Ruff lint errors

* Address MyPy errors

* feat: Add support for Kubernetes auth in Vault

* Add dagster-postgres as dependency to code locations

* Update lockfile

* fix: Add gcp_gcs key to resources for edxorg sensor

* Add missing "defs" directories for single-project dg execution

* feat: Allow OIDC Vault login when not using GH auth

* Fix broken dependencies for single-project launching

* Fix mongo package install

* fix: Use proper auth mount when not specified for GH auth