refactor: Switch to usin pyiceberg and returning a LazyFrame for dbt tables
By using the direct Polars method with a '.collect' it automatically reads all of the
table data into memory, even if much of it never gets used. This refactors the
`get_dbt_model_as_dataframe` method to use a lazy dataframe so that Polars can do
optimization of the operations before loading the resulting data into memory.